application.yml 수정
- profile 중에 local에서 기존에 local, jwt-local로 구분하여 사용하였는데
profile local에 대해서 jwt.active 값을 추가하여 true, false로 구분해서 사용하는 것으로 변경함
