+++ src/main/java/com/example/villagerservice/config/security/LocalSecurityConfig.java
@@ -1,11 +1,20 @@
 package com.example.villagerservice.config.security;
 
 import com.example.villagerservice.common.jwt.JwtTokenProvider;
+import com.example.villagerservice.config.redis.RedisRepository;
+import com.example.villagerservice.config.security.filters.CustomAuthenticationFilter;
+import com.example.villagerservice.config.security.filters.JwtAuthenticationFilter;
 import com.example.villagerservice.config.security.filters.LocalAuthenticationFilter;
+import com.example.villagerservice.config.security.handler.CustomFailureHandler;
+import com.example.villagerservice.config.security.handler.CustomSuccessHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.core.env.Environment;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
@@ -17,10 +26,14 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 @Configuration
 @EnableWebSecurity
 @RequiredArgsConstructor
-@Profile(value = {"jwt-local"})
+@Profile(value = {"local"})
 public class LocalSecurityConfig {
 
     private final JwtTokenProvider jwtTokenProvider;
+    private final RedisTemplate<String, Object> redisTemplate;
+    private final RedisRepository redisRepository;
+    private final Environment env;
+
     @Bean
     public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
@@ -28,6 +41,10 @@ public class LocalSecurityConfig {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+
+        AuthenticationManager authenticationManager = authenticationManager(http.getSharedObject(AuthenticationConfiguration.class));
+        CustomAuthenticationFilter authenticationFilter = createCustomAuthenticationFilter(authenticationManager);
+
         http
                 .headers()
                 .frameOptions()
@@ -36,7 +53,7 @@ public class LocalSecurityConfig {
 
         http
                 .authorizeRequests()
-                .antMatchers("/api/v1/auth/**", "/h2-console/**")
+                .antMatchers("/api/v1/auth/**", "/h2-console/**", "/docs/**")
                 .permitAll()
                 .anyRequest()
                 .authenticated();
@@ -49,9 +66,28 @@ public class LocalSecurityConfig {
                 .and()
                 .formLogin().disable()
                 .httpBasic().disable()
-                .addFilterBefore(new LocalAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class)
+                .addFilter(authenticationFilter)
         ;
 
+        if(Boolean.parseBoolean(env.getProperty("jwt.active"))) {
+            http.addFilterBefore(new JwtAuthenticationFilter(jwtTokenProvider), UsernamePasswordAuthenticationFilter.class);
+        } else {
+            http.addFilterBefore(new LocalAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
+        }
+
         return http.build();
+    }
+
+    private CustomAuthenticationFilter createCustomAuthenticationFilter(AuthenticationManager authenticationManager) {
+        CustomAuthenticationFilter authenticationFilter = new CustomAuthenticationFilter(authenticationManager);
+        authenticationFilter.setFilterProcessesUrl("/api/v1/auth/login");
+        authenticationFilter.setAuthenticationFailureHandler(new CustomFailureHandler());
+        authenticationFilter.setAuthenticationSuccessHandler(new CustomSuccessHandler(jwtTokenProvider, redisTemplate, redisRepository));
+        return authenticationFilter;
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration authConfiguration) throws Exception {
+        return authConfiguration.getAuthenticationManager();
     }
 }
